net: hns3: Modify return statement.
driver inclusion
category: bugfix
bugzilla: NA
CVE: NA

In function hns3_nic_bd_numa(), return type as Unsigned,
however the return value as Signed.
It would cause unexcept errors in some case.
This patch modify return statement to ensuring the return
value match.

Fix tag: d6a78e3c ("unify maybe_stop_tx for TSO and non-TSO case")

Feature or Bugfix:Bugfix

+++ drivers/net/ethernet/hisilicon/hns3/hns3_enet.c
@@ -1212,7 +1212,7 @@ static unsigned int hns3_nic_bd_num(struct sk_buff *skb)
 
 	/* if the total len is within the max bd limit */
 	if (likely(skb->len <= HNS3_MAX_BD_SIZE))
-		return skb_shinfo(skb)->nr_frags + 1;
+		return (unsigned int)(skb_shinfo(skb)->nr_frags + 1);
 
 	bd_num = hns3_tx_bd_count(skb_headlen(skb));
